Refactor and reorganized model files
