+++ rampp2p/models/__init__.py
@@ -1,14 +1,8 @@
-from .ad import *
-from .currency import *
-from .feedback import *
-from .order import *
-from .payment import *
-from .peer import *
-from .status import *
-from .appeal import *
-from .contract import *
-from .transaction import *
-from .market_rate import *
-from .arbiter import *
-from .models_utils import *
-from .models_slackbot import *
+from .model_peer import *
+from .model_arbiter import *
+from .model_ad import *
+from .model_order import *
+from .model_currency import *
+from .model_payment import *
+from .model_misc import *
+from .model_slackbot import *
