+++ rampp2p/models/model_misc.py
@@ -2,6 +2,18 @@ from django.db import models
 from datetime import date
 from django.conf import settings
 
+class MarketPrice(models.Model):
+    currency = models.CharField(max_length=10, unique=True)
+    price = models.DecimalField(max_digits=18, decimal_places=2, default=0)
+    created_at = models.DateTimeField(auto_now_add=True)
+    modified_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.currency
+    
+    class Meta:
+        ordering = ['currency']
+
 class AppVersion(models.Model):
     PLATFORM_CHOICES = [
         ('ios', 'iOS'),
