+++ rampp2p/views/views_order.py
@@ -300,7 +300,7 @@ class OrderViewSet(viewsets.GenericViewSet):
             self._check_create_permissions(wallet_hash, ad.id)
 
             # query market price for ad fiat currency
-            market_price = models.MarketRate.objects.filter(currency=ad.fiat_currency.symbol)
+            market_price = models.MarketPrice.objects.filter(currency=ad.fiat_currency.symbol)
             if market_price.exists():
                 market_price = market_price.first()
             else:
